extract a class and some methods from get commit class
